sửa lại code để giao diện đổ đk sản phẩm theo nhà sản xuất

+++ src/main/java/shop/dongho/repository/CustomerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import shop.dongho.model.Customer;
 
-public interface CustomerRepository extends PagingAndSortingRepository<Customer, Long> {
+public interface CustomerRepository extends PagingAndSortingRepository<Customer, Integer> {
     Page<Customer> findAllByNameContaining(String s, Pageable pageable);
 }
